Move georeference status into assertion
