+++ src/test/java/ca/gc/aafc/collection/api/entities/GeoreferenceAssertionCRUDIT.java
@@ -13,6 +13,7 @@ import javax.inject.Inject;
 import org.junit.jupiter.api.Test;
 
 import ca.gc.aafc.collection.api.CollectionModuleBaseIT;
+import ca.gc.aafc.collection.api.entities.GeoreferenceAssertion.GeoreferenceVerificationStatus;
 import ca.gc.aafc.collection.api.testsupport.factories.GeoreferenceAssertionFactory;
 import ca.gc.aafc.dina.testsupport.DatabaseSupportService;
 
@@ -22,6 +23,8 @@ public class GeoreferenceAssertionCRUDIT extends CollectionModuleBaseIT {
   private DatabaseSupportService dbService;
 
   private static final LocalDate testGeoreferencedDate = LocalDate.now();
+  private static final GeoreferenceVerificationStatus georeferenceVerificationStatus = GeoreferenceVerificationStatus.GEOREFERENCING_NOT_POSSIBLE;
+
 
   private List<UUID> agentIdentifiers = List.of(UUID.randomUUID(), UUID.randomUUID());
 
@@ -43,6 +46,7 @@ public class GeoreferenceAssertionCRUDIT extends CollectionModuleBaseIT {
         .dwcDecimalLongitude(45.01)
         .dwcCoordinateUncertaintyInMeters(10)
         .dwcGeoreferencedDate(testGeoreferencedDate)
+        .dwcGeoreferenceVerificationStatus(georeferenceVerificationStatus)
         .build();
     dbService.save(geoReferenceAssertion);
 
@@ -52,6 +56,7 @@ public class GeoreferenceAssertionCRUDIT extends CollectionModuleBaseIT {
     assertEquals(45.01, fetchedGeoreferenceAssertion.getDwcDecimalLongitude());
     assertEquals(10, fetchedGeoreferenceAssertion.getDwcCoordinateUncertaintyInMeters());
     assertEquals(testGeoreferencedDate, fetchedGeoreferenceAssertion.getDwcGeoreferencedDate());
+    assertEquals(georeferenceVerificationStatus, fetchedGeoreferenceAssertion.getDwcGeoreferenceVerificationStatus());
     assertNotNull(fetchedGeoreferenceAssertion.getCreatedOn());
   }
 }
